+++ jdaviz/utils.py
@@ -15,28 +15,52 @@ class SnackbarQueue:
     '''
     def __init__(self):
         self.queue = deque()
+        # track whether we're showing a loading message which won't clear by timeout,
+        # but instead requires another message with msg.loading = False to clear
+        self.loading = False
+        # track whether this is the first message - we'll increase the timeout for that
+        # to give time for the app to load.
         self.first = True
 
     def put(self, state, msg):
-        if len(self.queue) == 0:
+        if msg.loading:
+            # immediately show the loading message indefinitely until cleared by a new message
+            # with loading=False (or overwritten by a new indefinite message with loading=True)
+            self.loading = True
             self._write_message(state, msg)
-
-        if not msg.loading:
-            self.queue.appendleft(msg)
+        elif self.loading:
+            # clear the loading state, immediately show this message, then re-enter the queue
+            self.loading = False
+            self._write_message(state, msg)
+        else:
+            self.queue.append(msg)
+            if len(self.queue) == 1:
+                self._write_message(state, msg)
 
     def close_current_message(self, state):
 
-        # turn off snackbar and remove corresponding
-        # message from  queue.
+        # turn off snackbar iteself
         state.snackbar['show'] = False
+
+        if self.loading:
+            # then we've been interrupted, so keep this item in the queue to show after
+            # loading is complete
+            return
+
         if len(self.queue) > 0:
-            _ = self.queue.pop()
+            # determine if the closed entry came from the queue (not an interrupt)
+            # in which case we should remove it from the queue.  We clear here instead
+            # of when creating the snackbar so that items that are interrupted
+            # (ie by a loading message) will reappear again at the top of the queue
+            # so they are not missed
+            msg = self.queue[0]
+            if msg.text == state.snackbar['text']:
+                _ = self.queue.popleft()
 
         # in case there are messages in the queue still,
         # display the next.
         if len(self.queue) > 0:
-            msg = self.queue.pop()
-            self.queue.append(msg)
+            msg = self.queue[0]
             self._write_message(state, msg)
 
     def _write_message(self, state, msg):
@@ -49,6 +73,8 @@ class SnackbarQueue:
         state.snackbar['show'] = True
 
         if msg.loading:
+            # do not create timeout - the message will be indefinite until
+            # cleared by another message
             return
 
         # timeout of the first message needs to be increased by a
@@ -63,6 +89,8 @@ class SnackbarQueue:
             timeout += 5000
             self.first = False
 
+        # create the timeout function which will close this message and
+        # show the next message if one has been added to the queue since
         def sleep_function(timeout):
             timeout_ = float(timeout) / 1000
             time.sleep(timeout_)